requirements.txt
Django==4.2.7
gunicorn==21.2.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.2.0

# HuggingFace API (no local ML)
huggingface-hub==0.28.1
cloudinary
django-cloudinary-storage
dj-database-url

